+++ src/guia-envia/services/guia-envia.service.ts
@@ -32,6 +32,7 @@ import {
   EditAddressGEDataResponse,
 } from '../guia-envia.interface';
 import {
+  formatAddressesGE,
   formatCreateAddressPayloadGE,
   formatCreateAddressResponseGE,
   formatCreateGuidePayloadGE,
@@ -255,7 +256,7 @@ export class GuiaEnviaService {
     }
   }
 
-  async deleteGEAddress(alias: string): Promise<DeleteAddressGEDataResponse> {
+  async deleteGEAddress(id: string): Promise<DeleteAddressGEDataResponse> {
     try {
       const apiKey = this.configService.guiaEnvia.apiKey!;
       const uri = this.configService.guiaEnvia.uri!;
@@ -267,22 +268,7 @@ export class GuiaEnviaService {
         throw new BadRequestException(GE_MISSING_URI_ERROR);
       }
 
-      const getUri = `${uri}${CREATE_ADDRESS_ENDPOINT_GE}?limit=100`;
-      const responseGet: AxiosResponse<ExtGetAllAddressesGEResponse, unknown> =
-        await axios.get(getUri, {
-          headers: {
-            Authorization: apiKey,
-          },
-        });
-      const data = responseGet?.data;
-      const addressToDelete = (data?.data ?? []).find(
-        (address) => address.alias === alias,
-      );
-      if (!addressToDelete) {
-        throw new BadRequestException(`Address with alias ${alias} not found`);
-      }
-
-      const deleteUrl = `${uri}${CREATE_ADDRESS_ENDPOINT_GE}?id=${addressToDelete.id}`;
+      const deleteUrl = `${uri}${CREATE_ADDRESS_ENDPOINT_GE}?id=${id}`;
       await axios.delete(deleteUrl, {
         headers: {
           Authorization: apiKey,
@@ -297,6 +283,7 @@ export class GuiaEnviaService {
       };
     } catch (error) {
       if (axios.isAxiosError(error)) {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
         const newError = error?.response?.data || error.message;
         console.log('axios error deleting address ge', newError);
         throw new BadRequestException(error?.response?.data || error.message);
@@ -310,7 +297,13 @@ export class GuiaEnviaService {
     }
   }
 
-  async getAddressesSavedGe(page?: string): Promise<GetAliasesGEDataResponse> {
+  async getAddressesSavedGe({
+    page,
+    aliasesOnly,
+  }: {
+    page?: string;
+    aliasesOnly?: boolean;
+  }): Promise<GetAliasesGEDataResponse> {
     try {
       const apiKey = this.configService.guiaEnvia.apiKey!;
       const uri = this.configService.guiaEnvia.uri!;
@@ -330,13 +323,29 @@ export class GuiaEnviaService {
           },
         });
       const data = response?.data;
-      const aliases = (data?.data ?? []).map((address) => address.alias);
+      if (aliasesOnly) {
+        const aliases = (data?.data ?? []).map((address) => address.alias);
+        return {
+          version: npmVersion,
+          message: null,
+          error: null,
+          data: {
+            aliases,
+            addresses: [],
+            page: data?.meta?.page ?? 1,
+            pages: data?.meta?.pages ?? 1,
+          },
+        };
+      }
+
+      const addressesTransformed = formatAddressesGE(data?.data ?? []);
       return {
         version: npmVersion,
         message: null,
         error: null,
         data: {
-          aliases,
+          aliases: [],
+          addresses: addressesTransformed,
           page: data?.meta?.page ?? 1,
           pages: data?.meta?.pages ?? 1,
         },